refactor(pkg): move intelligence to internal/
- Move pkg/intelligence/ → internal/intelligence/
- Update all import paths to maintain functionality
- Follows Level 2 → internal strategy from design
- Build verification passes

Part 2/4 of pkg/ restructure for architecture compliance
