+++ internal/intelligence/behavior/circuit_breaker.go
