+++ internal/intelligence/behavior/constants.go
