+++ internal/intelligence/behavior/engine_test.go
