+++ internal/intelligence/behavior/integration_test.go
