+++ internal/intelligence/behavior/pattern_loader.go
