+++ internal/intelligence/behavior/prediction_integration_test.go
