+++ internal/intelligence/behavior/prediction_system_test.go
