+++ internal/intelligence/behavior/predictor.go
