+++ internal/intelligence/behavior/predictor_test.go
