+++ internal/intelligence/behavior/relationships.go
