+++ internal/intelligence/deployment_processor.go
