+++ internal/intelligence/safety_scorer.go
