+++ internal/intelligence/safety_scorer_test.go
