+++ internal/intelligence/types.go
