+++ internal/intelligence/types_test.go
